Update proc-create-credential.adoc (#696)
* Update proc-create-credential.adoc

Added final statement regarding the creation of a credential. At least one of three fields is mandatory, but it's also possible to create one in the web UI without specifying. The reason is because it's specified to the user creating it.

* Update proc-create-credential.adoc

Fixed the wording as per Robert's comment about awkwardness and removing the superfluous sentence.