Title updates aap 17967 (#1384)
* Update ref-operator-ocp-version.adoc (#1)

* Update ref-operator-ocp-version.adoc

changed the compatible platform from 4.9 later to 4.10-4.13, matching the life cycle doc and being the correct info

* Update ref-operator-ocp-version.adoc

Additional resources were not in fact being used

* Update master.adoc

Updated title as per Donna's comment on Nov. 7, 2023 in the *AAP-17967 comments

* Update master.adoc

Updated title as per Donna's comment on Nov. 7, 2023 in the AAP-17967 comments

* Update master.adoc

Updated title as per Donna's comment on Nov. 7, 2023 in the AAP-17967 comments

* Update master.adoc

Updated title as per Donna's comment on Nov. 7, 2023 in the AAP-17967 comments

* Update master.adoc

Updated title as per Donna's comment on Nov. 7, 2023 in the AAP-17967 comments

* Update master.adoc

Updated title as per Donna's comment on Nov. 7, 2023 in the AAP-17967 comments

* Update master.adoc

Updated title as per Donna's comment on Nov. 7, 2023 in the AAP-17967 comments

* Update master.adoc

Updated title as per Donna's comment on Nov. 7, 2023 in the AAP-17967 comments

* Update master.adoc

Updated title as per Donna's comment on Nov. 7, 2023 in the AAP-17967 comments

* Update master.adoc

Updated title as per Donna's comment on Nov. 7, 2023 in the AAP-17967 comments

* Update master.adoc

Updated title as per Donna's comment on Nov. 7, 2023 in the AAP-17967 comments

* Update master.adoc

Updated title as per Donna's comment on Nov. 7, 2023 in the AAP-17967 comments

* Update master.adoc

Updated title as per Donna's comment on Nov. 7, 2023 in the AAP-17967 comments

* Update master.adoc

Updated title as per Donna's comment on Nov. 7, 2023 in the AAP-17967 comments

* Update master.adoc

Updated title as per Donna's comment on Nov. 7, 2023 in the AAP-17967 comments

* Update master.adoc

Updated title as per Donna's comment on Nov. 7, 2023 in the AAP-17967 comments

* Update ref-operator-ocp-version.adoc

removed the tag